+++ backend/KotobaApi/Srs/Scheduling/IFsrsScheduler.cs
@@ -7,14 +7,14 @@ public sealed record FsrsScheduleRequest(
     DateTimeOffset? LastReviewedAt,
     DateTimeOffset ReviewedAt,
     FsrsRating Rating,
-    double? DesiredRetention = null);
+    float? DesiredRetention = null);
 
 public sealed record FsrsScheduleResult(
     Fsrs7MemoryState State,
-    double ElapsedDays,
-    double? RetrievabilityBeforeReview,
-    double DesiredRetention,
-    double IntervalDays,
+    float ElapsedDays,
+    float? RetrievabilityBeforeReview,
+    float DesiredRetention,
+    float IntervalDays,
     DateTimeOffset DueAt,
     bool WasFirstReview);
 
@@ -22,6 +22,6 @@ public interface IFsrsScheduler
 {
     string AlgorithmId { get; }
     FsrsScheduleResult Schedule(FsrsScheduleRequest request);
-    double Retrievability(Fsrs7MemoryState state, TimeSpan elapsed);
-    double IntervalAtRetention(Fsrs7MemoryState state, double desiredRetention);
+    float Retrievability(Fsrs7MemoryState state, TimeSpan elapsed);
+    float IntervalAtRetention(Fsrs7MemoryState state, float desiredRetention);
 }
